refactor:Change directory structure
