+++ src/content/styled/panel.ts
