+++ src/content/styled/tab.ts
